+++ src/main/java/laddergame/domain/ladder/RandomBooleanGenerator.java
@@ -1,4 +1,4 @@
-package laddergame.domain.rung;
+package laddergame.domain.ladder;
 
 import laddergame.util.BooleanGenerator;
