+++ src/test/java/laddergame/domain/ladder/RungTest.java
@@ -1,4 +1,4 @@
-package laddergame.domain.rung;
+package laddergame.domain.ladder;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
